Refatoring activate payment
